Commit with misates

+++ WorkOrderCore/Services/TransactionService.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using WorkOrderCore.Infrastructure.Helpers;
 using WorkOrderCore.Infrastructure.Persistence.DataContext;
 using WorkOrderCore.Model;
 
@@ -18,7 +19,7 @@ namespace WorkOrderCore.Services
         Task<JobCardsTranasctions> AddTransaction(JobCardsTranasctions model);
         Task<bool> UpdateTransaction(RiderChangeStatusViewModel model);
         Task<bool> CheckDuplicatePrioritySequence(int jobCardId, short prioritySequence);
-
+        Task<JobCardsTranasctions> EditTransaction(JobCardsTranasctions model);
         Task<List<AssignTransaction>> AssignedTransactionsByEmployeeId();
 
     }
@@ -74,10 +75,20 @@ namespace WorkOrderCore.Services
         {
             try
             {
-
+                short lastPriorityNumber = 0;
+                var jobPriority = await GetTransactionPriority(model.JobCardId);
+                if (jobPriority != null)
+                {
+                    lastPriorityNumber = (short)(jobPriority.PrioritySequence.Value + 1);
+                }
+                else
+                {
+                    lastPriorityNumber = 1;
+                }
                 model.CreatedDate = DateTime.Now;
                 model.UpdatedDate = DateTime.Now;
                 model.CreatedBy = LoginUserid;
+                model.PrioritySequence = lastPriorityNumber;
                 _context.JobCardsTranasctions.Add(model);
                 _context.SaveChanges();
 
@@ -91,7 +102,29 @@ namespace WorkOrderCore.Services
                     _context.AssignTransaction.Add(details);
                     _context.SaveChanges();
                 }
+                return model;
+            }
+            catch (Exception ex)
+            {
+                throw;
+            }
 
+        }
+
+        private async Task<JobCardsTranasctions> GetTransactionPriority(int jobCardId)
+        {
+             return await _context.JobCardsTranasctions.Where(c => c.JobCardId == jobCardId).OrderByDescending(c => c.CreatedDate).FirstOrDefaultAsync();
+        }
+
+        public async Task<JobCardsTranasctions> EditTransaction(JobCardsTranasctions model)
+        {
+            try
+            {
+                model.UpdatedBy = LoginUserid;
+                model.UpdatedDate = DateTime.Now;
+                model.CreatedBy = LoginUserid;
+                _context.JobCardsTranasctions.Update(model);
+                _context.SaveChanges();
                 return model;
             }
             catch (Exception ex)
@@ -105,31 +138,33 @@ namespace WorkOrderCore.Services
         {
             JobCardsTranasctions result = await _context.JobCardsTranasctions.Where(v => v.Id == model.Id).FirstOrDefaultAsync();
             result.JobCardsTranasctionsStatus = model.JobCardsTranasctionsStatus;
+            result.JobCardsTranasctionsClosedAt = DateTime.Now;
             result.UpdatedBy = LoginUserid;
             result.UpdatedDate = DateTime.Now;
             _context.JobCardsTranasctions.Update(result);
             await _context.SaveChangesAsync();
-
-            AssignTransaction removeDetails = new AssignTransaction
+            if (result.JobCardsTranasctionsStatus == "Close")
             {
-                TransactionId = result.Id,
-                EmployeeId = result.EmployeeId
-            };
-            _context.AssignTransaction.Remove(removeDetails);
-            await _context.SaveChangesAsync();
-
-            var getNextEmployee = await _context.JobCardsTranasctions.Where(c => c.JobActivityId == result.JobActivityId && c.PrioritySequence == result.PrioritySequence + 1).FirstOrDefaultAsync();
-            if (getNextEmployee != null)
-            {
-                AssignTransaction details = new AssignTransaction
+                AssignTransaction removeDetails = new AssignTransaction
                 {
                     TransactionId = result.Id,
-                    EmployeeId = getNextEmployee.EmployeeId
+                    EmployeeId = result.EmployeeId
                 };
-                _context.AssignTransaction.Add(details);
+                _context.AssignTransaction.Remove(removeDetails);
                 await _context.SaveChangesAsync();
-            }
 
+                var getNextEmployee = await _context.JobCardsTranasctions.Where(c => c.JobCardId == result.JobCardId && c.PrioritySequence == result.PrioritySequence + 1).FirstOrDefaultAsync();
+                if (getNextEmployee != null)
+                {
+                    AssignTransaction details = new AssignTransaction
+                    {
+                        TransactionId = result.Id,
+                        EmployeeId = getNextEmployee.EmployeeId
+                    };
+                    _context.AssignTransaction.Add(details);
+                    await _context.SaveChangesAsync();
+                }
+            }
             return true;
         }
